Move wagtail.admin.edit_handlers to wagtail.admin.panels
