+++ docs/topics/snippets.rst
@@ -17,7 +17,7 @@ Here's an example snippet model:
 
   from django.db import models
 
-  from wagtail.admin.edit_handlers import FieldPanel
+  from wagtail.admin.panels import FieldPanel
   from wagtail.snippets.models import register_snippet
 
   ...
