+++ wagtail/admin/checks.py
@@ -104,7 +104,7 @@ def inline_panel_model_panels_check(app_configs, **kwargs):
 
 def check_panels_in_model(cls, context="model"):
     """Check panels configuration uses `panels` when `edit_handler` not in use."""
-    from wagtail.admin.edit_handlers import BaseCompositeEditHandler, InlinePanel
+    from wagtail.admin.panels import BaseCompositeEditHandler, InlinePanel
     from wagtail.models import Page
 
     errors = []
