+++ wagtail/admin/forms/workflows.py
@@ -5,8 +5,8 @@ from django.utils.translation import gettext as _
 from django.utils.translation import gettext_lazy as __
 
 from wagtail.admin import widgets
-from wagtail.admin.edit_handlers import FieldPanel, InlinePanel, ObjectList
 from wagtail.admin.forms import WagtailAdminModelForm
+from wagtail.admin.panels import FieldPanel, InlinePanel, ObjectList
 from wagtail.admin.widgets.workflows import AdminTaskChooser
 from wagtail.coreutils import get_model_string
 from wagtail.models import Page, Task, Workflow, WorkflowPage
