+++ wagtail/contrib/forms/models.py
@@ -12,8 +12,8 @@ from django.utils.formats import date_format
 from django.utils.text import slugify
 from django.utils.translation import gettext_lazy as _
 
-from wagtail.admin.edit_handlers import FieldPanel
 from wagtail.admin.mail import send_mail
+from wagtail.admin.panels import FieldPanel
 from wagtail.contrib.forms.utils import get_field_clean_name
 from wagtail.models import Orderable, Page
