+++ wagtail/contrib/settings/views.py
@@ -9,7 +9,7 @@ from django.utils.text import capfirst
 from django.utils.translation import gettext as _
 
 from wagtail.admin import messages
-from wagtail.admin.edit_handlers import (
+from wagtail.admin.panels import (
     ObjectList,
     TabbedInterface,
     extract_panel_definitions_from_model_class,
