+++ wagtail/documents/edit_handlers.py
@@ -1,13 +1,13 @@
 from warnings import warn
 
-from wagtail.admin.edit_handlers import FieldPanel
+from wagtail.admin.panels import FieldPanel
 from wagtail.utils.deprecation import RemovedInWagtail219Warning
 
 
 class DocumentChooserPanel(FieldPanel):
     def __init__(self, *args, **kwargs):
         warn(
-            "wagtail.documents.edit_handlers.DocumentChooserPanel is obsolete and should be replaced by wagtail.admin.edit_handlers.FieldPanel",
+            "wagtail.documents.edit_handlers.DocumentChooserPanel is obsolete and should be replaced by wagtail.admin.panels.FieldPanel",
             category=RemovedInWagtail219Warning,
             stacklevel=2,
         )
